FIX predict.py and TEST github CI (linter+integration)
