--- .github/workflows/python-codestyle.yml
@@ -1,75 +1,0 @@
-name: Check codestyle
-
-on:
-  push:
-    branches: [dev]
-  pull_request:
-    branches: [dev]
-
-env:
-  NODE_TLS_REJECT_UNAUTHORIZED: 0
-
-jobs:
-  build:
-    runs-on: [self-hosted, Windows]
-    strategy:
-      # For debbug purposes
-      fail-fast: true
-      matrix:
-        python-version: ["3.10.4"]
-    steps:
-    - uses: actions/checkout@v3
-    - name: Checlout conda
-      run: conda deactivate
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v3
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Cache pip
-      uses: actions/cache@v2
-      with:
-          path: ~/.cache/pip
-          key: ${{ runner.os }}-pip-${{ hashFiles('**/requirements.txt') }}
-          restore-keys: |
-            ${{ runner.os }}-pip-
-    - name: Install dependencies
-      run: |
-        python --version
-        python -m pip install -- upgrade pip
-        pip install mypy pylint flake8 pytest
-    - name: Analysing the code with pylint
-      run: |
-        pylint $(git ls-files '*.py') --disable=E0401
-    - name: Format with black
-      run: |
-        black --check .
-
-    # - name: Run dvc workflow
-    #   run: |
-    #     conda activate mlops-course
-    #     # "conda run --no-capture-output -n ${CONDA_ENV}
-    #     dvc pull
-    #     dvc repro
-    # - name: Run unit tests
-    #   run: |
-    #     conda actiate mlops-course
-    #     pytest --disable-warnings
-        
-    # - name: "Check type hints by mypy"
-      # run: |
-      #   mypy --ignore-missing-imports .gitignore
-      #   flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-    # - name: Analysing the code with pylint
-    #   run: |
-    #     pylint $(git ls-files '*.py')
-
-# steps:
-# - uses: actions/checkout@v3
-# - name: Install poetry
-#   run: pipx install poetry
-# - uses: actions/setup-python@v3
-#   with:
-#     python-version: '3.9'
-#     cache: 'poetry'
-# - run: poetry install
-# - run: poetry run pytest
