+++ data/raw.dvc
@@ -1,0 +1,5 @@
+outs:
+- md5: 740993cae74002c2587001a4a5306b07.dir
+  size: 22905874
+  nfiles: 9
+  path: raw
