+++ models/metrics.json
@@ -1,11 +1,4 @@
 {
-    "test": {
-        "MSE": 0.02733,
-        "MAE": 0.11543,
-        "MAE_median": 0.11598,
-        "MAE_std": 0.00059,
-        "MAE_eval": 0.11657
-    },
     "train": {
         "MAE": 0.12
     }
