+++ reports/figures/.gitignore
@@ -1,0 +1,4 @@
+/exploratory/*
+/importance/*
+!/exploratory/.gitkeep
+!/importance/.gitkeep
