+++ tox.ini
@@ -1,3 +1,3 @@
 [flake8]
 max-line-length = 88
-ignore = E203
+ignore = [E203, E999]
